move mode reducer into shared
