+++ src/admin/containers/member_page.js
@@ -6,7 +6,7 @@ const title_case = require('app/to_title_case')
 const { convertPounds } = require('app/monies')
 const { fetch_member, deactivate_member, reactivate_member, update_member } =
   require('../redux/modules/member.js')
-const { toggle_member_mode } = require('../redux/modules/mode.js')
+const { toggle_member_mode } = require('../../shared/redux/modules/mode.js')
 const { switch_charge_type } = require('../redux/modules/charge_form.js')
 const { add_payment, remove_payment } = require('../redux/modules/payments.js')
